Updating save to the new Toml API (#217)
### What's done:
- new API of ktoml (0.2.7)
- changed Int to Long as Toml does not support Int type

+++ save-plugins/fix-plugin/src/commonMain/kotlin/org/cqfn/save/plugins/fix/FixPluginConfig.kt
@@ -27,7 +27,7 @@ import kotlinx.serialization.UseSerializers
 @Serializable
 data class FixPluginConfig(
     val execFlags: String? = null,
-    val batchSize: Int? = null,
+    val batchSize: Long? = null,
     val batchSeparator: String? = null,
     val resourceNameTestSuffix: String? = null,
     val resourceNameExpectedSuffix: String? = null,
